use CodeWF.Tools;

+++ src/SocketTesting/SocketDto/Response/ResponseServiceInfo.cs
@@ -53,7 +53,7 @@ public class ResponseServiceInfo : INetObject
     ///     通信对象时间戳起始年份，比如：2023，表示2023年1月1号开始计算时间戳，后面的时间戳都以这个字段计算为准，精确到0.1s，即100ms，主要用于节约网络对象传输大小
     /// </summary>
     [Key(7)]
-    public byte TimestampStartYear { get; set; }
+    public int TimestampStartYear { get; set; }
 
     /// <summary>
     ///     最后更新时间
